Add `cliTool`
This CLI tool is an example in how to use the classic unix lib to make a
terminal program. I love it, I would love it more if it used ncurses.
For now it does NOTHING, and it is great to have an example when we need
to implement a new of these.