Add `pthreads`
Learning the POSIX threads API is very fun most of the time, though I
think this implementation is incomplete. The real fun of the project was
to make a thread be the user interface and the other one be the
"database" where N user inputs where saved. This is fun actually. But
sadly incomplete.

I am commiting in order to keep track of it. Keep it organized.